requirements.txt
# Snowflake
snowflake-connector-python[pandas]==3.12.3
python-dotenv==1.0.1

# Data processing
pandas==2.2.3
requests==2.32.3

# AIS WebSocket
websocket-client==1.8.0

# dbt
dbt-snowflake==1.9.0

# API
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1

# Testing
pytest==8.3.4
httpx==0.28.1
